Add tests and builder
