+++ Tests/Given_an_existing_set_with_one_element_of_data.cs
@@ -1,4 +1,4 @@
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Linq;
 using Xunit;
 using WallyMathieu.Collections;
@@ -6,30 +6,30 @@ using With;
 
 namespace Tests
 {
-    public class Given_an_existing_set_with_one_element
+    public class Given_an_existing_set_with_one_element_of_data
     {
         ImmutableHashSet<int> existing = new[] { 1 }.ToImmutableHashSet();
 
         [Fact]
         public void When_there_is_an_incoming_element() =>
-            Assert.Equal(existing.SymmetricDiff(Plus(2, existing).ToHashSet()),
-                new DifferenceBuilder<int>
+            Assert.Equal(Differences.Of().Existing(existing).Incoming(Plus(2, existing).ToHashSet()),
+                new DataDifferenceBuilder<int>
                 {
-                    OnlyInRight = { 2 }
+                    ToBeAdded = { 2 }
                 }.Build());
 
         [Fact]
         public void When_there_an_empty_incoming_list() =>
-            Assert.Equal(existing.SymmetricDiff(new int[0].ToHashSet()),
-                new DifferenceBuilder<int>
+            Assert.Equal(Differences.Of().Existing(existing).Incoming(new int[0].ToHashSet()),
+                new DataDifferenceBuilder<int>
                 {
-                    OnlyInLeft = { 1 }
+                    ToBeDeleted = { 1 }
                 }.Build());
 
         [Fact]
         public void When_incoming_is_the_same_as_existing() =>
-            Assert.Equal(existing.SymmetricDiff(existing),
-                new DifferenceBuilder<int>().Build());
+            Assert.Equal(Differences.Of().Existing(existing).Incoming(existing),
+                new DataDifferenceBuilder<int>().Build());
 
         private ImmutableHashSet<T> Plus<T>(T v, ImmutableHashSet<T> arr) =>
             arr.ToHashSet().Tap(l => l.Add(v)).ToImmutableHashSet();
