+++ WallyMathieu.Collections/Abstractions/IDataDifference.cs
@@ -1,14 +1,29 @@
 ﻿using System.Collections.Generic;
 
-namespace WallyMathieu.Collections
+namespace WallyMathieu.Collections.Abstractions
 {
+    /// <summary>
+    /// Represents a symmetric difference between two collections where one side repressents deletions and the other additions
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    public interface IDataDifference<out T>
+    {
+        /// <summary>
+        /// Items with keys only in the right collection.
+        /// </summary>
+        IReadOnlyCollection<T> ToBeDeleted { get; }
+        /// <summary>
+        /// Items with keys only in the left collection.
+        /// </summary>
+        IReadOnlyCollection<T> ToBeAdded { get; }
+    }
     /// <summary>
     /// Represents a symmetric difference between two dictionaries or maps where the values have potentially different type 
     /// </summary>
     /// <typeparam name="TKey"></typeparam>
     /// <typeparam name="TIncoming"></typeparam>
     /// <typeparam name="TExisting"></typeparam>
-    public interface IIncomingDifference<TKey, TIncoming, TExisting>
+    public interface IDataDifference<TKey, TIncoming, TExisting>
     {
         /// <summary>
         /// Items with keys only in the right collection.
@@ -21,6 +36,6 @@ namespace WallyMathieu.Collections
         /// <summary>
         /// Items with same keys only both collections.
         /// </summary>
-        IReadOnlyCollection<IKeyIncomingIntersection<TKey, TIncoming, TExisting>> Intersection { get; }
+        IReadOnlyCollection<IDataIntersection<TKey, TIncoming, TExisting>> Intersection { get; }
     }
 }
