+++ WallyMathieu.Collections/SetExtensions.cs
@@ -27,22 +27,6 @@ namespace WallyMathieu.Collections
 
             return new SymmetricDifference<T>(onlyInRight:onlyInRight, onlyInLeft:onlyInLeft);
         }
-
-        /// <summary>
-        /// Compare two sets.
-        /// The comparison is the equivalent of doing <code>right.Except(left)</code> and <code>left.Except(right)</code>
-        /// </summary>
-        /// <example>
-        /// <code>
-        /// var comparison = db.PublishedProducts.ToHashSet().SymmetricDiff(body.ProductIds.ToHashSet());
-        /// var onlyInRight = comparison.OnlyInRight;
-        /// var onlyInLeft = comparison.OnlyInLeft;
-        /// </code>
-        /// </example>
-        public static IIncomingDifferenceBuilder<T> Existing<T>(this ISet<T> existing)
-        {
-            return new IncomingDifferenceBuilder<T>(existing);
-        }
     }
    
 }
